% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bz_data.r
\name{get_amcat_data}
\alias{get_amcat_data}
\title{Download articles from AmCAT}
\usage{
get_amcat_data(
  conn,
  project,
  set,
  headline_col = "headline",
  medium_col = "medium",
  clean = T
)
}
\arguments{
\item{conn}{An amcatr connection}

\item{project}{AmCAT project}

\item{set}{AmCAT articleset}

\item{headline_col}{The name of the headline column}

\item{medium_col}{The name of the medium column}

\item{clean}{There is some mess in the data causing weird tokens. This does some rough regex cleaning}
}
\value{
A data.frame
}
\description{
This is a simple wrapper for downloading the full-text articles from AmCAT. Note that the headline and medium column
might be different for more recent versions of AmCAT. This function renames them to prepare the data for \code{\link{create_bz_topics_data}}.
}
