% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{plot_topic_quality}
\alias{plot_topic_quality}
\title{Run stm::topicQuality for current topicbrowser data}
\usage{
plot_topic_quality()
}
\value{
Nothing, just plots
}
\description{
Run stm::topicQuality for current topicbrowser data
}
