% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_topicbrowser}
\alias{run_topicbrowser}
\title{Run the Shiny Application}
\usage{
run_topicbrowser(bz_data, token_auth = F, port = 6171, topic_names = NULL, ...)
}
\arguments{
\item{bz_data}{the data, prepared with create_bz_topics_data()}

\item{token_auth}{If True, read the .valid_tokens file (create with \code{\link{set_tokens}}) to authenticate tokens. 
The token should be passed to the app with the ?token URL parameter}

\item{port}{The port}
}
\description{
Run the Shiny Application
}
