% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{set_tokens}
\alias{set_tokens}
\title{Create .valid_tokens file}
\usage{
set_tokens(tokens)
}
\arguments{
\item{tokens}{}
}
\value{
safety
}
\description{
Creates the .valid_tokens file. Previous versions will be overwritten 
(so only the most recently created tokens are valid)
}
