% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_topicbrowser}
\alias{run_topicbrowser}
\title{Run the Shiny Application}
\usage{
run_topicbrowser(token_auth = F, port = 6171, topic_selection = NULL, ...)
}
\arguments{
\item{token_auth}{If True, read the .valid_tokens file (create with \code{\link{set_tokens}}) to authenticate tokens. 
The token should be passed to the app with the ?token URL parameter}

\item{port}{The port}

\item{topic_selection}{Optionally, the indices of topics to show}
}
\description{
This uses the data most recently created with \code{\link{create_bz_topics_data}}.
In case you have (accidentally) overwritten data that you want to use, see \code{\link{use_data_backup}}
}
\details{
For some data, such as topic names, the app uses persistent storage shared by all users. 
You can thus set the topic names within the app.
}
